requirements.txt
python-vlc>=3.0.0
Pillow>=8.0.0